refactor: centralize provider config field metadata
