+++ sdk/packages/core/src/index.ts
@@ -441,13 +441,9 @@ export {
 	deleteLocalProvider,
 	ensureCustomProvidersLoaded,
 	getLocalProviderModels,
-	getProviderConfigFields,
 	listLocalProviders,
 	loginLocalProvider,
 	normalizeOAuthProvider,
-	type ProviderConfigFieldKey,
-	type ProviderConfigFieldRequirement,
-	type ProviderConfigFields,
 	refreshProviderModelsFromSource,
 	resolveLocalClineAuthToken,
 	saveLocalProviderOAuthCredentials,
@@ -455,6 +451,12 @@ export {
 	type UpdateLocalProviderRequest,
 	updateLocalProvider,
 } from "./services/providers/local-provider-service";
+export {
+	getProviderConfigFields,
+	type ProviderConfigFieldKey,
+	type ProviderConfigFieldRequirement,
+	type ProviderConfigFields,
+} from "./services/providers/provider-config-fields";
 export {
 	type MigrateLegacyProviderSettingsOptions,
 	type MigrateLegacyProviderSettingsResult,
