+++ sdk/packages/core/src/services/providers/local-provider-service.ts
@@ -3,7 +3,6 @@ import {
 	type AddProviderActionRequest,
 	getClineEnvironmentConfig,
 	type ITelemetryService,
-	isOAuthProviderId,
 	type OAuthProviderId,
 	type ProviderCapability,
 	type ProviderListItem,
@@ -800,121 +799,4 @@ export function resolveLocalClineAuthToken(
 ): string | undefined {
 	const token = settings?.auth?.accessToken?.trim() || settings?.apiKey?.trim();
 	return token && token.length > 0 ? token : undefined;
-}
-
-// --- Provider configuration fields (UI projection) -------------------------
-
-export type ProviderConfigFieldKey =
-	| "apiKey"
-	| "baseUrl"
-	| "awsRegion"
-	| "awsProfile";
-
-export interface ProviderConfigFieldRequirement {
-	defaultValue?: string;
-	label?: string;
-	note?: string;
-	placeholder?: string;
-	optional?: boolean;
-}
-
-export interface ProviderConfigFields {
-	providerId: string;
-	authMethod: "api-key" | "oauth" | "local";
-	fields: Partial<
-		Record<ProviderConfigFieldKey, ProviderConfigFieldRequirement>
-	>;
-	/** Optional description shown above the fields (e.g. AWS region auto-fill hint). */
-	description?: string;
-}
-
-const EDITABLE_BASE_URL_PROVIDER_IDS = new Set([
-	"ollama",
-	"lmstudio",
-	"litellm",
-	"openai-compatible",
-]);
-
-function shouldExposeBaseUrlField(
-	providerId: string,
-	collection: LlmsModels.ModelCollection | undefined,
-): boolean {
-	if (!collection?.provider.baseUrl) return false;
-	if (collection.provider.source !== "system") return true;
-	return EDITABLE_BASE_URL_PROVIDER_IDS.has(providerId);
-}
-
-/**
- * Project a provider into the inputs a configure-dialog should render.
- *
- * No fields are marked "required". `llms` no longer pre-flights credentials,
- * so a missing API key surfaces as the provider's own auth error rather than
- * a synthetic SDK failure. UIs may still require fields client-side if they
- * want, but the runtime does not.
- *
- * - OAuth providers (`cline`, `oca`, `openai-codex`) return `authMethod:
- *   "oauth"` with no fields; the configure UI should route to the OAuth
- *   login flow instead.
- * - Local auth providers return `authMethod: "local"` with no fields. The
- *   configure UI should show provider-specific local readiness instead.
- * - All other providers return `apiKey`. Built-in local/proxy-style providers
- *   with user-supplied endpoints, plus user-added providers with saved
- *   endpoints, also return a pre-filled `baseUrl` field.
- *
- * Returns the same fallback shape for unknown providers (single `apiKey`
- * input, no default base URL) so callers can render a reasonable configure
- * dialog without per-id branches.
- */
-export function getProviderConfigFields(
-	providerId: string,
-): ProviderConfigFields {
-	const id = LlmsModels.normalizeProviderId(providerId);
-	if (isOAuthProviderId(id)) {
-		return { providerId: id, authMethod: "oauth", fields: {} };
-	}
-
-	const collection = LlmsModels.MODEL_COLLECTIONS_BY_PROVIDER_ID[id];
-	if (collection?.provider.capabilities?.includes("local-auth")) {
-		return { providerId: id, authMethod: "local", fields: {} };
-	}
-
-	if (id === "bedrock") {
-		return {
-			providerId: id,
-			authMethod: "api-key",
-			description:
-				"AWS region is required for Bedrock. It can be auto-filled from AWS_REGION, AWS_DEFAULT_REGION, or ~/.aws/config.",
-			fields: {
-				awsRegion: {
-					label: "AWS Region",
-					placeholder: "us-east-1",
-				},
-				apiKey: {
-					label: "AWS Bedrock API Key (optional)",
-					placeholder: "Leave blank to use AWS profile/default chain",
-					optional: true,
-				},
-				awsProfile: {
-					label: "AWS Profile Name (optional)",
-					placeholder: "default",
-					optional: true,
-				},
-			},
-		};
-	}
-
-	const defaultBaseUrl = collection?.provider.baseUrl;
-	const fields: ProviderConfigFields["fields"] = {
-		apiKey:
-			id === "ollama"
-				? {
-						note: "Keep empty if no API key for local inference.",
-					}
-				: {},
-	};
-	if (shouldExposeBaseUrlField(id, collection)) {
-		fields.baseUrl = { defaultValue: defaultBaseUrl };
-	}
-
-	return { providerId: id, authMethod: "api-key", fields };
 }
